# src/controllers/appointment.controller.js
const { appointmentService } = require('../services')
const catchAsync = require('../utils/catchAsync')
const { StatusCodes } = require('http-status-codes')


const createAppointment = catchAsync(async (req, res) => {
  const { title, description, startDate, endDate, author } = req.body
  const appointment = await appointmentService.createAppointment({
    title,
    description,
    startDate,
    endDate,
    author
  })

  res.status(StatusCodes.CREATED).send({
    appointment
  })
})

const getAppointments = catchAsync(async (req, res) => {
  const appointments = await appointmentService.getAppointments()
  res.status(StatusCodes.OK).send({
    appointments
  })
})

const getAppointmentById = catchAsync(async (req, res) => {
  const id = req.params.id
  const appointment = await appointmentService.getAppointmentById(id)
  res.status(StatusCodes.OK).send({
    appointment
  })
})

const updateAppointment = catchAsync(async (req, res) => {
  const id = req.params.id
  const updateBody = req.body
  const appointment = await appointmentService.updateAppointment(id, updateBody)

  res.status(StatusCodes.OK).send({
    appointment
  })
})

const deleteAppointment = catchAsync(async (req, res) => {
  const id = req.params.id
  const appointment = await appointmentService.deleteAppointment(id)
  res.status(StatusCodes.OK).send(appointment)
})

module.exports = {
  createAppointment,
  getAppointments,
  getAppointmentById,
  updateAppointment,
  deleteAppointment
}